refactor: Move linear search implementation to blocks-cs-search crate
