+++ crates/cs/blocks-cs-search/src/linear.rs
