--- src/algorithms/mod.rs
@@ -1,3 +1,0 @@
-mod algorithms;
-
-pub use algorithms::linear_search;
